backend/requirements.txt
django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.1
gunicorn>=21.2
whitenoise>=6.6
django-filter>=23.5
